backend/requirements.txt
anyio==3.5.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.72.0
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.5
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
